+++ CHANGELOG.md
@@ -5,6 +5,7 @@ Recommended Gamevault Server Version: `v12.1.1`
 ### Changes
 
 - Desktop shortcuts start the game via GameVault. So changed executable or parameters are updated immediately as well.
+- Option to create a shortcut upon clicking install. Remembers the last input as default.
 - Implemented Jumplist to taskbar icon.
 - Allow user to specify rows displayed of installed games.
 - Notifications about download/extraction events now also display the game box image